+++ packages/component-driver-mui-x-v5-test/package.json
@@ -8,8 +8,8 @@
     "src"
   ],
   "scripts": {
-    "start": "PORT=5125 TSC_COMPILE_ON_ERROR=true react-scripts start",
-    "build:ui": "react-scripts build",
+    "start": "vite --force",
+    "build:ui": "vite build",
     "postbuild": "node ../../scripts/postBuild.mjs",
     "test:dom": "jest",
     "test:e2e": "playwright test",
@@ -62,10 +62,11 @@
     "@types/node": "^22.15.30",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "@vitejs/plugin-react": "^4.5.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^5.8.3"
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5"
   },
   "eslintConfig": {
     "extends": [
