+++ packages/component-driver-mui-x-v5-test/src/examples/datagridpro/index.ts
@@ -1,8 +1,9 @@
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
 
-import { basicDataGridProExample, basicDataGridProTestSuite } from './BasicDataGridPro.examples';
+import { basicDataGridProUIExample } from './BasicDataGridPro.examples';
+import { basicDataGridProExample, basicDataGridProTestSuite } from './BasicDataGridPro.suite';
 
-export { basicDataGridProExample, basicDataGridProTestSuite };
+export { basicDataGridProUIExample, basicDataGridProExample, basicDataGridProTestSuite };
 export const dataGridProExamples: IExampleUnit<ScenePart, JSX.Element>[] = [
   basicDataGridProExample,
 ] satisfies IExampleUnit<ScenePart, JSX.Element>[];
